qt: estimate the time left from blocks, not from the progress percentage
"Estimated time left until synced" read "unknown" through most of a sync. It
divided the remaining progress by how much the percentage had moved over the
last ~8 minutes, and verification progress is weighted by each block's
transaction load: on this chain the early blocks are nearly empty, so the
percentage crawls for most of the sync and then lurches at the end. Too small a
delta, no estimate — precisely while the user is waiting and wants one.

Measure the rate in blocks instead. Heights are integers that always advance,
so blocks_left x (elapsed / blocks_done) yields a figure within seconds of
starting and stays steady. Average over a 60-second window, falling back to the
whole history while the sync has only just begun. If no block at all was taken
on within the window, keep saying "unknown" rather than invent a number.

+++ src/qt/modaloverlay.h
@@ -56,7 +56,10 @@ private:
         not derived from the consensus constant (600s here vs ~34s actual). */
     GUIUtil::HeaderSyncEstimator m_header_estimator;
     QDateTime bestHeaderDate;
-    QVector<QPair<qint64, double> > blockProcessTime;
+    //! (wall-clock ms, block height) samples, newest first: the sync rate is
+    //! measured in blocks, which always advance, rather than in the progress
+    //! percentage, which can sit still for most of a sync.
+    QVector<QPair<qint64, int> > blockProcessTime;
     bool layerIsVisible;
     bool userClosed;
     QPropertyAnimation m_animation;
